==> Forecasting_WNV/Notebooks-Master/01B_Data_Cleansing_Traps_Master.ipynb <==
# 01B Data Cleansing Traps

## Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle
%matplotlib inline

## Read Data

df = pd.read_csv('../west_nile/west_nile/input/train.csv')

## Read in codes_df and temperature_df

with open('../Cleansed_Data/codes_df.pkl', 'rb') as f:
    codes_df = pickle.load(f)

with open('../Cleansed_Data/new_weather_df.pkl', 'rb') as f:
    new_weather_df = pickle.load(f)    

## Calculate Total Mosquitos Per Trap

cols = [col for col in df.columns if col not in ['NumMosquitos', 'WnvPresent']]
df = df.groupby(cols, as_index=False).sum()
df['WnvPresent'] = (df.WnvPresent>0).astype(int)

## Join codes_df and new_weather_df

df = df.join(codes_df, on='Date', rsuffix="_code_")
df = df.join(new_weather_df, on='Date')

## Extract Month from Date

df['Month'] = df.Date.str[5:7]

## One-hot encoding

df = pd.get_dummies(df, columns=['Month', 'Species'])

## Drop Unnecessary Columns

cols_drop = ['Address','Block','Street','Trap','AddressNumberAndStreet','AddressAccuracy','Date']
df = df.drop(labels=cols_drop, axis=1)

## Split X and y Variables

target = 'WnvPresent'
X = df.drop(target, 1)
y = df[target]

## Export X and y Variables

with open('../Cleansed_Data/X.pkl','wb') as f:
    pickle.dump(X, f)
    
with open('../Cleansed_Data/y.pkl','wb') as f:
    pickle.dump(y, f)
