==> Forecasting_WNV/Notebooks-Master/02_Preprocessing_Master.ipynb <==
# 02 Preprocessing

## Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle
%matplotlib inline

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn.preprocessing import StandardScaler

## Read in X and y Variables

with open('../Cleansed_Data/X.pkl', 'rb') as f:
    X = pickle.load(f)
    
with open('../Cleansed_Data/y.pkl', 'rb') as f:
    y = pickle.load(f)

## Drop NumMosquitos from X

X = X.drop('NumMosquitos', axis=1)

### Train Test Split

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 42, stratify=y)

### Use Standard Scalar

ss = StandardScaler()
X_train_ss = pd.DataFrame(ss.fit_transform(X_train), columns=X_train.columns, index=X_train.index)
X_test_ss = pd.DataFrame(ss.fit_transform(X_test), columns=X_test.columns, index=X_test.index)


with open('../Objects/ss.pkl','wb') as f:
    pickle.dump(ss, f)

## Export Training and Test Data

with open('../Cleansed_Data/X_train.pkl','wb') as f:
    pickle.dump(X_train, f)
    
with open('../Cleansed_Data/X_test.pkl','wb') as f:
    pickle.dump(X_test, f)

with open('../Cleansed_Data/X_train_ss.pkl','wb') as f:
    pickle.dump(X_train_ss, f)
    
with open('../Cleansed_Data/X_test_ss.pkl','wb') as f:
    pickle.dump(X_test_ss, f)
    
with open('../Cleansed_Data/y_train.pkl','wb') as f:
    pickle.dump(y_train, f)
    
with open('../Cleansed_Data/y_test.pkl','wb') as f:
    pickle.dump(y_test, f)
